# swagger/openapi.yaml
openapi: 3.0.0
info:
  title: The race cars
  contact:
    email: loi.tra@9thwonder.com
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/
  - url: https://virtserver.swaggerhub.com/annguyen9w/hapi-starter/1.0.0
    description: SwaggerHub API Auto Mocking
paths:
...
  /drivers:
    get:
      tags:
        - Driver
      summary: Get all driver
      operationId: getDrivers
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DriverResponse'
        # '400':
        #   $ref: '#/components/responses/badrequest'
    post:
      tags:
        - Driver
      summary: Add a new driver to the system
      operationId: addDriver
      requestBody:
        $ref: '#/components/requestBodies/Driver'
      responses:
        '200':
          $ref: '#/components/responses/successwithid'
        '400':
          $ref: '#/components/responses/badrequest'
  /drivers/{driverId}:
    get:
      tags:
        - Driver
      summary: Find driver by ID
      operationId: getDriverById
      parameters:
        - name: driverId
          in: path
          description: ID of driver to return
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverResponse'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
    put:
      tags:
        - Driver
      summary: Update an existing driver
      operationId: updateDriver
      parameters:
        - name: driverId
          in: path
          description: Driver id need to be updated
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        $ref: '#/components/requestBodies/Driver'
      responses:
        '200':
          $ref: '#/components/responses/success'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
    delete:
      tags:
        - Driver
      summary: Delete a driver
      operationId: deleteDriver
      parameters:
        - name: driverId
          in: path
          description: Driver id to delete
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          $ref: '#/components/responses/success'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
  /drivers/{driverId}/results:
    get:
      tags:
        - Driver
      summary: All race results for that driver
      operationId: getRaceResultsByDriverId
      parameters:
        - name: driverId
          in: path
          description: ID of driver to return
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RaceResult'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
...
  /races:
    get:
      tags:
        - Race
      summary: Get all race
      operationId: getRaces
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Race'
        # '400':
        #   $ref: '#/components/responses/badrequest'
    post:
      tags:
        - Race
      summary: Add a new race to the system
      operationId: addRace
      requestBody:
        $ref: '#/components/requestBodies/CreateRace'
      responses:
        '200':
          $ref: '#/components/responses/successwithid'
        '400':
          $ref: '#/components/responses/badrequest'
  /races/{raceId}:
    get:
      tags:
        - Race
      summary: Find race by ID
      operationId: getRaceById
      parameters:
        - name: raceId
          in: path
          description: ID of race to return
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Race'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
    put:
      tags:
        - Race
      summary: Update an existing race
      operationId: updateRace
      parameters:
        - name: raceId
          in: path
          description: Race id need to be updated
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        $ref: '#/components/requestBodies/UpdateRace'
      responses:
        '200':
          $ref: '#/components/responses/success'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
    delete:
      tags:
        - Race
      summary: Delete a race
      operationId: deleteRace
      parameters:
        - name: raceId
          in: path
          description: Race id to delete
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          $ref: '#/components/responses/success'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
  /races/{raceId}/results:
    get:
      tags:
        - Race
      summary: All race results for that race
      operationId: getRaceResultsByRaceId
      parameters:
        - name: raceId
          in: path
          description: ID of race to return
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RaceResult'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
    post:
      tags:
        - Race
      summary: Add race results for that race
      operationId: addRaceResults
      parameters:
        - name: raceId
          in: path
          description: ID of race to add
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        $ref: '#/components/requestBodies/AddRaceResult'
      responses:
        '200':
          $ref: '#/components/responses/success'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
  /race-results/{raceResultId}:
    put:
      tags:
        - Race Result
      summary: Update an existing race result
      operationId: updateRaceResult
      parameters:
        - name: raceResultId
          in: path
          description: Race result id need to be updated
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        $ref: '#/components/requestBodies/RaceResult'
      responses:
        '200':
          $ref: '#/components/responses/success'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
    delete:
      tags:
        - Race Result
      summary: Delete a race result
      operationId: deleteRaceResult
      parameters:
        - name: raceResultId
          in: path
          description: Race result id to delete
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          $ref: '#/components/responses/success'
        '400':
          $ref: '#/components/responses/badrequest'
        '404':
          $ref: '#/components/responses/notfound'
components:
  schemas:
    UUID:
      type: string
      format: uuid
      example: 40e6215d-b5c6-4896-987c-f30f3678f608
    ClassRequestBody:
      required:
        - name
      type: object
      properties:
        name:
          type: string
          example: LM GTE AM
    Class:
      allOf:
        - required:
            - id
          type: object
          properties:
            id:
              $ref: '#/components/schemas/UUID'
        - $ref: '#/components/schemas/ClassRequestBody'
    AddressRequestBody:
      required:
        - city
        - country
        - state
        - zipcode
      type: object
      properties:
        street:
          type: string
          example: 1797 Colonial Drive
        street2:
          type: string
          example: 411 Patterson Street
        city:
          type: string
          example: Houston
        state:
          type: string
          example: Texas
        zipcode:
          type: string
          example: '77002'
        country:
          type: string
          example: United States
    Address:
      allOf:
        - required:
            - id
          type: object
          properties:
            id:
              $ref: '#/components/schemas/UUID'
        - $ref: '#/components/schemas/AddressRequestBody'
    DriverRequestBody:
      required:
        - firstName
        - lastName
        - nationality
      type: object
      properties:
        firstName:
          type: string
          example: Nguyen
        lastName:
          type: string
          example: An
        nationality:
          type: string
          enum:
            - USA
            - Viet Nam
        homeAddress:
          $ref: '#/components/schemas/UUID'
        managementAddress:
          $ref: '#/components/schemas/UUID'
    Driver:
      required:
        - firstName
        - id
        - lastName
        - nationality
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        firstName:
          type: string
          example: Nguyen
        lastName:
          type: string
          example: An
        nationality:
          type: string
          enum:
            - USA
            - Viet Nam
        homeAddress:
          $ref: '#/components/schemas/Address'
        managementAddress:
          $ref: '#/components/schemas/Address'
    DriverResponse:
      allOf:
        - $ref: '#/components/schemas/Driver'
        - required:
            - teams
          type: object
          properties:
            teams:
              type: array
              items:
                $ref: '#/components/schemas/Team'
...
    Team:
      required:
        - cars
        - id
        - name
        - nationality
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: 7 TOYOTA GAZOO RACING
        nationality:
          type: string
          enum:
            - USA
            - Viet Nam
        businessAddress:
          $ref: '#/components/schemas/Address'
...
    Car:
      required:
        - class
        - make
        - model
        - team
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        make:
          type: string
          example: Ferrari
        model:
          type: string
          example: 488 GTE-LM
        class:
          $ref: '#/components/schemas/Class'
...
    RaceRequestBody:
      required:
        - name
      type: object
      properties:
        name:
          type: string
          example: Le Mans
    Race:
      required:
        - name
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: Le Mans
    RaceResultRequestBody:
      required:
        - car
        - class
        - driver
        - raceNumber
        - startPosition
      type: object
      properties:
        car:
          $ref: '#/components/schemas/UUID'
        driver:
          $ref: '#/components/schemas/UUID'
        class:
          $ref: '#/components/schemas/UUID'
        raceNumber:
          type: string
          example: '200'
        startPosition:
          minimum: 1
          type: integer
          example: 2
        finishPosition:
          minimum: 1
          type: integer
          example: 3
    RaceResult:
      required:
        - car
        - class
        - driver
        - id
        - race
        - raceNumber
        - startPosition
      type: object
      properties:
        id:
          type: string
          format: uuid
        race:
          $ref: '#/components/schemas/Race'
        car:
          $ref: '#/components/schemas/Car'
        driver:
          $ref: '#/components/schemas/Driver'
        class:
          $ref: '#/components/schemas/Class'
        raceNumber:
          type: string
          example: '200'
        startPosition:
          minimum: 1
          type: integer
          example: 2
        finishPosition:
          minimum: 1
          type: integer
          example: 3
  requestBodies:
...
    Driver:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DriverRequestBody'
      required: true
...
    CreateRace:
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/RaceRequestBody'
              - type: object
                properties:
                  raceResults:
                    type: array
                    items:
                      $ref: '#/components/schemas/RaceResultRequestBody'
      required: true
    UpdateRace:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RaceRequestBody'
      required: true
    AddRaceResult:
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/RaceResultRequestBody'
      required: true
    RaceResult:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RaceResultRequestBody'
      required: true
  responses:
    success:
      description: Successful operation
      content:
        text/plain:
          schema:
            type: string
            example: success
    successwithid:
      description: Successful operation
      content:
        application/json:
          schema:
            type: object
            properties:
              id: 
                $ref: '#/components/schemas/UUID'
              status:
                type: string
                default: success
            required:
              - id
              - status
    badrequest:
      description: Invalid input
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode: 
                type: integer
                default: 400
              error:
                type: string
                default: Bad Request
              message:
                type: string
                default: Invalid request input
            required:
              - statusCode
              - error
              - message
    notfound:
      description: Item not found
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode: 
                type: integer
                default: 404
              error:
                type: string
                default: Not Found
              message:
                type: string
                default: Not Found
            required:
              - statusCode
              - error
              - message